🐛  Fix: 이제 Response객체의 total필드가 Elasticsearch의 전체 문서 hits수를 반환합니다.

+++ src/main/java/com/likelion/backendplus4/talkpick/backend/search/presentation/controller/dto/response/NewsSearchResponseList.java
@@ -16,5 +16,5 @@ import lombok.RequiredArgsConstructor;
 @Builder
 public class NewsSearchResponseList {
 	private final List<NewsSearchResponse> newsSearchResponseList;
-	private final int total;
+	private final Long total;
 }
